Use Synthetic bean for producing OpenTelemetry
This will make the bean production even more
resilient to timing issues
