+++ extensions/opentelemetry/runtime/src/main/java/io/quarkus/opentelemetry/runtime/exporter/otlp/RemoveableLateBoundBatchSpanProcessor.java
@@ -1,12 +1,10 @@
 package io.quarkus.opentelemetry.runtime.exporter.otlp;
 
-import io.quarkus.opentelemetry.runtime.OpenTelemetryProducer;
+import io.quarkus.opentelemetry.runtime.AutoConfiguredOpenTelemetrySdkBuilderCustomizer.TracerProviderCustomizer;
 
 /**
- * The only point in having this class is to allow {@link OpenTelemetryProducer}
+ * The only point in having this class is to allow {@link TracerProviderCustomizer}
  * to easily ignore the configured {@link LateBoundBatchSpanProcessor}.
- * <p>
- * In the future when {@link OpenTelemetryProducer} is replaced by a synthetic bean, this class will no longer be necessary
  */
 public final class RemoveableLateBoundBatchSpanProcessor extends LateBoundBatchSpanProcessor {
